requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
stripe==8.5.0